Fix unpack method when there is only one element in the first tree level, but this is a binary, not a direcoty -like in consul-
Reviewers: andres, dbarranco

Maniphest Tasks: T24743

Differential Revision: http://phabricator.bitnami.com:8080/D24372

+++ index.js
@@ -174,7 +174,7 @@ function unpack(file, destination, options) {
     const tmpDest = _getUniqueFile(`${destination.replace(/\/*$/, '')}.tmp`);
     _unpack(file, tmpDest, options);
     const contents = fs.readdirSync(tmpDest);
-    if (contents.length === 1) {
+    if (contents.length === 1 && fs.statSync(path.join(tmpDest, contents[0])).isDirectory()) {
       nfile.rename(path.join(tmpDest, contents[0]), destination);
       nfile.delete(tmpDest);
     } else {
